Remove redundant numeric conversion
Agent-Logs-Url: https://github.com/hussainshahzad350/phone_shop_pos/sessions/4991256f-f3de-4d4f-adf9-c5653529e937

+++ lib/modules/reports/services/operations_workflow_service.dart
@@ -415,8 +415,7 @@ class OperationsWorkflowService with BaseRepositoryGuard {
         final salePaymentMethod = PaymentMethod.normalizeNullable(
           saleRows.first['payment_method'] as String?,
         );
-        final newTotal =
-            (currentTotal - returnAmount).clamp(0.0, currentTotal).toDouble();
+        final newTotal = (currentTotal - returnAmount).clamp(0.0, currentTotal);
         final newPaid = currentPaid.clamp(0.0, newTotal).toDouble();
         final refundedPaidAmount =
             (currentPaid - newPaid).clamp(0, currentPaid).toDouble();
